requirements.txt
pandas
numpy
scikit-learn
matplotlib
statsmodels
streamlit
python-dateutil
catboost